feature: User profile edit, change password, logout
